[BC] moved useLazyLoadQuery into useQuery.ts and removed useQueryFetcher
