+++ src/usePreloadedQuery.ts
@@ -1,17 +1,18 @@
-import { useState, useEffect } from 'react';
+import { useEffect } from 'react';
 import { OperationType } from 'relay-runtime';
 import { RenderProps, LoadQuery } from './RelayHooksType';
+import { useForceUpdate } from './useForceUpdate';
 import { useRelayEnvironment } from './useRelayEnvironment';
 
 export const usePreloadedQuery = <TOperationType extends OperationType = OperationType>(
     loadQuery: LoadQuery,
 ): RenderProps<TOperationType> => {
-    const [, forceUpdate] = useState();
+    const forceUpdate = useForceUpdate();
     const environment = useRelayEnvironment();
 
     useEffect(() => {
-        const dispose = loadQuery.subscribe(forceUpdate);
-        return (): void => dispose();
+        return loadQuery.subscribe(forceUpdate);
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [loadQuery]);
 
     return loadQuery.getValue(environment) as RenderProps<TOperationType>;
